Break up the monolthic store
