+++ src/components/pages/ConfirmPurchase.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script>
-  import { mapGetters, mapState } from 'vuex';
+  import {mapGetters, mapState} from 'vuex';
   import Artist from '../Artist';
   import GalleryEdition from '../GalleryEdition';
   import ConfirmPurchaseButton from '../ui-controls/ConfirmPurchaseButton';
@@ -35,7 +35,7 @@
       ConfirmPurchaseButton
     },
     computed: {
-      ...mapGetters([
+      ...mapGetters('assets', [
         'firstAssetForEdition'
       ]),
       edition: function () {
